refactore: reorganize apps.drs package and export public entities
